28oct

+++ C03/ex03/ft_strncat.c
@@ -6,14 +6,14 @@
 /*   By: sclaude <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/24 10:06:14 by sclaude           #+#    #+#             */
-/*   Updated: 2020/10/26 14:13:25 by sclaude          ###   ########.fr       */
+/*   Updated: 2020/10/27 17:23:10 by sclaude          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 char	*ft_strncat(char *dest, char *src, unsigned int nb)
 {
-	int i;
-	int j;
+	int				i;
+	unsigned int	j;
 
 	i = 0;
 	j = 0;
@@ -21,7 +21,7 @@ char	*ft_strncat(char *dest, char *src, unsigned int nb)
 	{
 		i++;
 	}
-	while ((unsigned int) j < nb && src[j] != '\0')
+	while (j < nb && src[j] != '\0')
 	{
 		dest[i] = src[j];
 		i++;
